fix(ci): improve test timing for expiring_data module
Real-world timing is always a bad idea when it comes to tests and CI.

In this case, we handle expiring data, and for now the best idea we
could come up with was to make the data expire after a short time span,
and wait until that has passed. Checking that the data is still valid
immediately after the creation, then wait for a long enough period of
time, then check again that the data has expired.

I assume that the switch to `cargo nextest` increases parallelization in
the CI runs to a degree where the first of these checks fails, because
the 3ms have passed already. I hope increasing this to 10ms fixes these
CI failures, because the test should still run in a timely fashion, and
we should not introduce any delays in the magnitude of 100ms or more,
because that makes the tests non-timely.