fix(storage): make long-lasting uploads such as recordings robust against database connection drops
Closes #1163

We're observing files being stored in the file storage, and at the same
time not being listed in the room's file list.

The cause appears to be that we acquire a database connection from
the pool first, and keeping it for further use after the upload to the
file. Connections sometimes get closed in the meantime, resulting in the
database entry not being written after the upload.

This change introduces one database connection being acquired from the
pool and handed back before the upload, and another after the upload
finished. That makes the procedure more robust against closed database
connections, and results in less resources being blocked during the
upload.